pyproject.toml
[project]
name = "notebooks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "b>=1.0.0",
    "glob2>=0.7",
    "jedi>=0.19.2",
    "kagglehub>=0.3.12",
    "keras>=3.9.2",
    "marimo[lsp]>=0.13.10",
    "matplotlib>=3.10.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "polars>=1.29.0",
    "scikit-learn>=1.6.1",
    "tensorflow>=2.19.0",
    "tqdm>=4.67.1",
]

[tool.marimo.experimental]
lsp = true

# Language server configuration
[tool.marimo.language_servers.pylsp]
enabled = true               # Enable/disable the Python language server
enable_mypy = true           # Type checking with mypy (enabled by default, if installed)
enable_ruff = true           # Linting with ruff (enabled by default, if installed)
enable_flake8 = false        # Linting with flake8
enable_pydocstyle = false    # Check docstring style
enable_pylint = false        # Linting with pylint
enable_pyflakes = false      # Syntax checking with pyflakes

# Diagnostics configuration
[tool.marimo.diagnostics]
enabled = true               # Show diagnostics in the editor
